+++ src/fuzzifino/model/sang_modes.jl
@@ -167,7 +167,7 @@ enables the Hermitian conjugate of a spherical mode.
 """
 function Base.adjoint(amd :: SAngModes)
     l2m = amd.l2m
-    amd1 = SAngModes(l2m, (l2, m2) -> (iseven((l2 + m2) ÷ 2) ? 1 : -1) * amd.get_comp(l2, -m2)')
+    amd1 = SAngModes(l2m, (l2, m2) -> (1.0im) ^ (-m2) * amd.get_comp(l2, -m2)')
     return amd1
 end
 
@@ -185,9 +185,7 @@ function Base.:*(amd1 :: SAngModes, amd2 :: SAngModes)
     l2m2 = amd2.l2m
     l2m = l2m1 + l2m2
     gc = ((l2, m2) -> sum(STerms[sum(STerms[sum(STerms[
-            (iseven((-l21 + l22 + m2) ÷ 2) ? 1 : -1) *
-            sqrt(l2 + 1) *
-            wigner3j(l21/2, l22/2, l2/2, m21/2, (m2-m21)/2, -m2/2) *
+            clebschgordan(l21/2, m21/2, l22/2, (m2-m21)/2, l2/2) *
             amd1.get_comp(l21, m21) * amd2.get_comp(l22, m2 - m21)
         for m21 = max(-l21, -l22 + m2) : 2 : min(l21, l22 + m2)])
         for l21 = max(l2m1 % 2, abs(l2 - l22)) : 2 : min(l2m1, l2 + l22)])
@@ -212,15 +210,18 @@ end
 
 Return the contraction of two angular modes 
 ```math 
-    ∑_{l}U_l∑_{m=-l}^l(-1)^{l+m}Φ_{1,l}Φ_{2,l(-m)}
+    ∑_{l}U_l∑_{m=-l}^l(-1)^mΦ_{1,lm}Φ_{2,l(-m)}
 ```
 where the list of ``U_l`` is given by a dictionary, or
 ```math 
-    U_{l₀}∑_{m=-l₀}^{l₀}(-1)^{l₀+m}Φ_{1,l₀m}Φ_{2,l₀(-m)}.
+    U_{l₀}∑_{m=-l₀}^{l₀}(-1)^mΦ_{1,l₀m}Φ_{2,l₀(-m)}.
 ```
 """
 function ContractMod(amd1 :: SAngModes, amd2 :: SAngModes, comps :: Dict)
-    return SimplifyTerms(sum(STerms[ U * amd1.get_comp(Int64(2 * l), m2) * amd2.get_comp(Int64(2 * l), -m2) * (iseven((Int64(2 * l) + m2) ÷ 2) ? 1 : -1) for (l, U) in comps for m2 = -Int64(2 * l) : 2 : Int64(2 * l)]))
+    return SimplifyTerms(sum(STerms[
+        U * amd1.get_comp(Int64(2 * l), m2) * amd2.get_comp(Int64(2 * l), -m2) * (1.0im) ^ (m2)
+        for (l, U) in comps for m2 = -Int64(2 * l) : 2 : Int64(2 * l)
+    ]))
 end
 ContractMod(amd1 :: SAngModes, amd2 :: SAngModes, l0 :: Number) = ContractMod(amd1, amd2, Dict([l0 => 1]))
 
@@ -255,7 +256,7 @@ end
 """
     GetFermionSMod(nm :: Int64, nf :: Int64, f :: Int64[ ; mom_incr :: Bool]) :: SAngModes
 
-returns the modes of fermion annihilation operator ``c_m``, with angular momentum ``s=(N_m-1)/2``
+returns the modes of fermion annihilation operator ``f_m``, with angular momentum ``s=(N_m-1)/2``. Its components are ``(f)_m=(-1)^m f_{-m}`` and its adjoint ``(f^†)_m=f^†_m``.
 
 # Arguments
 
@@ -266,7 +267,7 @@ returns the modes of fermion annihilation operator ``c_m``, with angular momentu
 """
 function GetFermionSMod(nm :: Int64, nf :: Int64, f :: Int64 ; mom_incr :: Bool = ObsMomIncr)
     if mom_incr
-        gc = (l2, m2) -> (l2 == nm - 1) ? STerms((-1) ^ ((nm - 1 - m2) ÷ 2), [0, f + nf * ((nm - 1 - m2) ÷ 2)]) : STerm[]
+        gc = (l2, m2) -> (l2 == nm - 1) ? STerms((1.0im) ^ m2, [0, f + nf * ((nm - 1 - m2) ÷ 2)]) : STerm[]
     else
         gc = (l2, m2) -> (l2 == nm - 1) ? STerms(1.0, [0, f + nf * ((m2 + nm - 1) ÷ 2)]) : STerm[]
     end
@@ -277,7 +278,7 @@ end
 """
     GetBosonSMod(nm :: Int64, nf :: Int64, f :: Int64[ ; mom_incr :: Bool]) :: SAngModes
 
-returns the modes of electron boson operator ``b_m``, with angular momentum ``s=(N_m-1)/2``
+returns the modes of electron boson operator ``b_m``, with angular momentum ``s=(N_m-1)/2``. Its components are ``(b)_m=(-1)^m b_{-m}`` and its adjoint ``(b^†)_m=b^†_m``.
 
 # Arguments
 
@@ -288,7 +289,7 @@ returns the modes of electron boson operator ``b_m``, with angular momentum ``s=
 """
 function GetBosonSMod(nm :: Int64, nf :: Int64, f :: Int64 ; mom_incr :: Bool = ObsMomIncr)
     if mom_incr
-        gc = (l2, m2) -> (l2 == nm - 1) ? STerms((-1) ^ ((nm - 1 - m2) ÷ 2), [0, -(f + nf * ((nm - 1 - m2) ÷ 2))]) : STerm[]
+        gc = (l2, m2) -> (l2 == nm - 1) ? STerms((1.0im) ^ m2, [0, -(f + nf * ((nm - 1 - m2) ÷ 2))]) : STerm[]
     else
         gc = (l2, m2) -> (l2 == nm - 1) ? STerms(1.0, [0, -(f + nf * ((m2 + nm - 1) ÷ 2))]) : STerm[]
     end
@@ -299,9 +300,9 @@ end
 """
     GetFerPairingSMod(nm :: Int64, nf :: Int64, mat :: Matrix{<:Number}[ ; mom_incr :: Bool]) :: SAngModes
 
-returns the modes of two fermions superposed in the rule of CG coefficients. 
+returns the modes of two fermions superposed in the rule of CG coefficients. Its adjoint is
 ```math 
-    Δ_{lm}=∑_{m_1m_2}δ_{m,m_1+m_2}⟨sm_1,sm_2|lm⟩c_{a,m_1}M_{ab}c_{b,m_2}
+    Δ^†_{lm}=∑_{m_1m_2}δ_{m,m_1+m_2}⟨sm_1,sm_2|lm⟩c^†_{a,m_1}M_{ab}c^†_{b,m_2}
 ```
 
 # Arguments
@@ -325,9 +326,9 @@ end
 """
     GetBosPairingSMod(nm :: Int64, nf :: Int64, mat :: Matrix{<:Number}[ ; mom_incr :: Bool]) :: SAngModes
 
-returns the modes of two bosons superposed in the rule of CG coefficients. 
+returns the modes of two bosons superposed in the rule of CG coefficients. Its adjoint is
 ```math 
-    Δ_{lm}=∑_{m_1m_2}δ_{m,m_1+m_2}⟨sm_1,sm_2|lm⟩b_{a,m_1}M_{ab}b_{b,m_2}
+    Δ^†_{lm}=∑_{m_1m_2}δ_{m,m_1+m_2}⟨sm_1,sm_2|lm⟩b^†_{a,m_1}M^†_{ab}b^†_{b,m_2}
 ```
 
 # Arguments
@@ -353,7 +354,7 @@ end
 
 returns the modes of electron creation and annihilation superposed in the rule of CG coefficients. 
 ```math 
-    n_{lm}=∑_{m_1m_2}δ_{m,-m_1+m_2}(-1)^{s+m_1}⟨s(-m_1),sm_2|lm⟩c^†_{m_1}c_{m_2}
+    n_{lm}=∑_{m_1m_2}δ_{m,-m_1+m_2}(-1)^{-m_2}⟨sm_1,s(-m_2)|lm⟩f^†_{m_1}f_{m_2}
 ```
 
 # Arguments
@@ -380,7 +381,7 @@ end
 
 returns the modes of boson creation and annihilation superposed in the rule of CG coefficients. 
 ```math 
-    n_{lm}=∑_{m_1m_2}δ_{m,-m_1+m_2}(-1)^{s+m_1}⟨s(-m_1),sm_2|lm⟩b^†_{m_1}b_{m_2}
+    n_{lm}=∑_{m_1m_2}δ_{m,-m_1+m_2}(-1)^{-m_2}⟨sm_1,s(-m_2)|lm⟩b^†_{m_1}b_{m_2}
 ```
 
 # Arguments
@@ -403,7 +404,6 @@ end
 
 """
     PadSAngModes(amd :: SAngModes, nofl :: Int64, nobl :: Int64)
-
 
 adds `nofl` fermionic and `nobl` bosonic empty orbitals to the left by shifting each orbital index, implemented as